Revert "Fix rendering inaccuracy in triangles"
This reverts commit ece3ed03856027896a988e03ef55cdaffa3609b6.

+++ src/helpers/pax_dh_mcr_unshaded.c
@@ -139,7 +139,7 @@ void paxmcr_tri_unshaded(bool odd_scanline, pax_buf_t *buf, pax_col_t color,
 			x_b += x0_x2_dx;
 		}
 		int delta = y * buf->width;
-		for (; y <= y_pre_2; y += 2) {
+		for (; y <= (int) y_pre_2; y += 2) {
 			// Plot the horizontal line.
 			float x_left, x_right;
 			if (x_a < x_b) {
